2004-09-23  Marcus Brinkmann  <marcus@g10code.de>
	* data-stream.c (stream_seek): Call ftello and return the current
	offset.
	* data.h (struct gpgme_data): Change type of data.mem.offset to
	off_t.
	* data.c (gpgme_data_seek): Check dh->cbs->seek callback, not read
	callback.  If SEEK_CUR, adjust the offset by the pending buffer
	size.  Clear pending buffer on success.

+++ gpgme/data.h
@@ -99,7 +99,7 @@ struct gpgme_data
       /* Allocated size of BUFFER.  */
       size_t size;
       size_t length;
-      size_t offset;
+      off_t offset;
     } mem;
 
     /* For gpgme_data_new_from_read_cb.  */
